day 3 part 2

+++ src/test/kotlin/Day03Test.kt
@@ -8,78 +8,78 @@ class Day03Test {
     @Test
     fun day03_part1_12345() {
         val actual = joltage("12345", 2)
-        assertEquals(45.toBigInteger(), actual)
+        assertEquals(45, actual)
     }
 
     @Test
     fun day03_part1_987654321111111() {
         val actual = joltage("987654321111111", 2)
-        assertEquals(98.toBigInteger(), actual)
+        assertEquals(98, actual)
     }
 
     @Test
     fun day03_part1_234234234234278() {
         val actual = joltage("234234234234278", 2)
-        assertEquals(78.toBigInteger(), actual)
+        assertEquals(78, actual)
     }
 
     @Test
     fun day03_part1_818181911112111() {
         val actual = joltage("818181911112111", 2)
-        assertEquals(92.toBigInteger(), actual)
+        assertEquals(92, actual)
     }
 
     @Test
     fun day03_part2_12345() {
         val actual = joltage("12345", 3)
-        assertEquals(345.toBigInteger(), actual)
+        assertEquals(345, actual)
     }
 
     @Test
     fun day03_part2_987654321111111() {
         val actual = joltage("987654321111111", 3)
-        assertEquals(987.toBigInteger(), actual)
+        assertEquals(987, actual)
     }
 
     @Test
     fun day03_part2_811111111111119() {
         val actual = joltage("811111111111119", 3)
-        assertEquals(819.toBigInteger(), actual)
+        assertEquals(819, actual)
     }
 
     @Test
     fun day03_part2_234234234234278() {
         val actual = joltage("234234234234278", 3)
-        assertEquals(478.toBigInteger(), actual)
+        assertEquals(478, actual)
     }
 
     @Test
     fun day03_part2_818181911112111() {
         val actual = joltage("818181911112111", 3)
-        assertEquals(921.toBigInteger(), actual)
+        assertEquals(921, actual)
     }
 
     @Test
     fun day03_part2_987654321111111_for_12() {
         val actual = joltage("987654321111111", 12)
-        assertEquals(987654321111.toBigInteger(), actual)
+        assertEquals(987654321111, actual)
     }
 
     @Test
     fun day03_part2_811111111111119_for_12() {
         val actual = joltage("811111111111119", 12)
-        assertEquals(811111111119.toBigInteger(), actual)
+        assertEquals(811111111119, actual)
     }
 
     @Test
     fun day03_part2_234234234234278_for_12() {
         val actual = joltage("234234234234278", 12)
-        assertEquals(434234234278.toBigInteger(), actual)
+        assertEquals(434234234278, actual)
     }
 
     @Test
     fun day03_part2_818181911112111_for_12() {
         val actual = joltage("818181911112111", 12)
-        assertEquals(888911112111.toBigInteger(), actual)
+        assertEquals(888911112111, actual)
     }
 }
